Got the lexer to be 100% coverage. Fixed many problems and bugs. Still need to add proper line tests and many tokens instead of 1 or 2.
Add support for binary numbers, including a corresponding option.
Renamed 'intrinsic' as 'native' which is the correct keyword in JavaScript.
Fixed the Message class so it supports the iomanip.
Fixed the support for string continuation (backslash and newline).
Fixed the read_identifier() function so it only accepts valid identifier characters.
Fixed the read_hex() so it handles lower and upper case letters separately.
Fixed the escape_sequence() which does not support \U by default, extended escape sequences must be turned on.
Fixed the support of an escape sequence for identifiers.
Fixed all the keywords that had to use 'return' instead of 'break'.
Fixed the fact that regular expression expect a string.
Fixed comment tests to support the '\0' character in them.
Fixed the operator order for match (and not match).
Removed the code in the input stream that would ignore the BOM because it has to be viewed as a space by the lexer.
Removed the necessity of not having strict turned on if extended escape sequences is on to make use of '\e' and octal sequences.
Removed all the extended statement from the keywords. That way we are safer as no one can use those as variable or function names.
Added a few missing space characters to the char_type() function.
Added an ungetc() in the lexer otherwise newlines could be counted multiple times.
Added all the EMCAScript version 5 reserved keywords (abstract, boolean, byte, char, debugger, double, export, final, float, long, protected, short, static, synchronized, throws, transient, volatile, yield).
Added support for Eiffel like function contracts with these new keywords: ensure, invariant, require, then.
Added a compare keyword (<=>) and a not match (!~) and smart match (~~).
Added a test to generate an error when a number is followed by a letter.
Added an error when an invalid character is found as is in the input stream.
Added support for identifiers to start with a backslash character.
Added support for inline regular expressions (without the `...` quotes).
Added options that were not in the pragma definitions.
Added a command to test Unicode characters in different ways.
Changed the rotate to '!<' and '!>' because the other way around it could be a < !b or a > !b.
Changed the &&=, ^^, ^^=, ||= operators so they are extended operators.
Changed the message char support to print a character, not a number.
Various clean ups.

+++ as2js/lib/message.cpp
@@ -224,7 +224,7 @@ Message& Message::operator << (String const& s)
 
 /** \brief Append a char value.
  *
- * This function appends a char value to the message.
+ * This function appends a character to the message.
  *
  * \param[in] v  A char value.
  *
@@ -232,7 +232,7 @@ Message& Message::operator << (String const& s)
  */
 Message& Message::operator << (char const v)
 {
-    static_cast<std::stringstream&>(*this) << static_cast<int>(v);
+    static_cast<std::stringstream&>(*this) << v;
     return *this;
 }
